Split local-cluster-slow into two
