+++ ci/test-local-cluster-slow-2.sh
@@ -1,0 +1,1 @@
+test-stable.sh
